+++ src/inject-babel-helpers.ts
@@ -1,19 +1,21 @@
-import * as dom5 from 'dom5';
-import {fs} from 'mz';
-import * as parse5 from 'parse5';
-import * as path from 'path';
+/**
+ * @license
+ * Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
 
+import {htmlTransform} from './html-transform';
 import {AsyncTransformStream, getFileContents} from './streams';
 
 import File = require('vinyl');
-
-const p = dom5.predicates;
-
-const scriptOrImport = p.OR(
-    p.hasTagName('script'),
-    p.AND(p.hasTagName('link'), p.hasSpaceSeparatedAttrValue('rel', 'import')));
-
-
 
 /**
  * When compiling to ES5 we need to inject Babel's helpers into a global so
@@ -35,29 +37,9 @@ export class BabelHelpersInjector extends AsyncTransformStream<File, File> {
       return file;
     }
     const contents = await getFileContents(file);
-    const document = parse5.parse(contents);
-
-    const babelHelpersFragment =
-        parse5.parseFragment('\n\n<script></script>\n\n');
-    dom5.setTextContent(
-        babelHelpersFragment.childNodes![1]!,
-        await fs.readFile(
-            path.join(__dirname, 'babel-helpers.min.js'), 'utf-8'));
-
-    const firstScriptOrImport = dom5.nodeWalk(document, scriptOrImport);
-    if (firstScriptOrImport) {
-      dom5.insertBefore(
-          firstScriptOrImport.parentNode!,
-          firstScriptOrImport,
-          babelHelpersFragment);
-    } else {
-      const head =
-          dom5.query(document, dom5.predicates.hasTagName('head')) || document;
-      dom5.append(head, babelHelpersFragment);
-    }
-
+    const transformed = htmlTransform(contents, {injectBabelHelpers: true});
     const newFile = file.clone();
-    newFile.contents = new Buffer(parse5.serialize(document), 'utf-8');
+    newFile.contents = new Buffer(transformed, 'utf-8');
     return newFile;
   }
 }