+++ src/test/util.ts
@@ -13,6 +13,7 @@
  */
 
 import File = require('vinyl');
+import {assert} from 'chai';
 
 export function getFlowingState(stream: NodeJS.ReadableStream): boolean {
   // Cast our streams to <any> so that we can check the flowing state.
@@ -41,3 +42,44 @@ export async function emittedFiles(
               .on('end', () => resolve(files))
               .on('error', (e: Error) => reject(e)));
 }
+
+/**
+ * Assert that two strings are equal after collapsing their whitespace.
+ */
+export const assertEqualIgnoringWhitespace =
+    (actual: string, expected: string) =>
+        assert.equal(collapseWhitespace(actual), collapseWhitespace(expected));
+
+/**
+ * Assert that two string maps are equal, where their values have had their
+ * whitespace collapsed.
+ */
+export const assertMapEqualIgnoringWhitespace =
+    (actual: Map<string, string>, expected: Map<string, string>) =>
+        assertMapEqual(
+            transformMapValues(actual, collapseWhitespace),
+            transformMapValues(expected, collapseWhitespace));
+
+/**
+ * Collapse all leading whitespace, trailing whitespace, and newlines. Very
+ * lossy, but good for loose comparison of HTML, JS, etc.
+ */
+const collapseWhitespace = (s: string) =>
+    s.replace(/^\s+/gm, '').replace(/\s+$/gm, '').replace(/\n/gm, '');
+
+/**
+ * Assert that two maps are equal. Note that early versions of chai's deepEqual
+ * will always return true, and while later ones will compare correctly, they do
+ * not produce very readable output compared to this approach.
+ */
+const assertMapEqual = <K, V>(actual: Map<K, V>, expected: Map<K, V>) =>
+    assert.deepEqual([...actual.entries()], [...expected.entries()]);
+
+/**
+ * Return a new map where all values have been transformed with the given
+ * function.
+ */
+const transformMapValues =
+    <K, V1, V2>(map: Map<K, V1>, transform: (val: V1) => V2): Map<K, V2> =>
+        new Map([...map.entries()].map(
+            ([key, val]): [K, V2] => [key, transform(val)]));