+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* ES to AMD module transformation is now supported by `getOptimizeStreams` and `htmlTransform`. Additionally:
+  * Will now inject an inline minified RequireJS AMD loader, and the Babel helpers.
+  * Phantom `<html>`, `<head>`, and `<body>` tags injected by parse5 are now removed.
+
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.3] - 2018-03-18